Updating everything with Anmol's suggestions
